+++ Tests/GRDBTests/DatabasePoolBackupTests.swift
@@ -8,15 +8,9 @@ import XCTest
 class DatabasePoolBackupTests: GRDBTestCase {
 
     func testBackup() throws {
-        #if GRDBCIPHER
-        // SQLCipher can't backup encrypted databases: skip this test
-        if dbConfiguration.passphrase != nil {
-            return
-        }
-        #endif
-        
-        let source = try makeDatabasePool(filename: "source.sqlite")
-        let destination = try makeDatabasePool(filename: "destination.sqlite")
+        // SQLCipher can't backup encrypted databases: use a pristine Configuration
+        let source = try makeDatabasePool(filename: "source.sqlite", configuration: Configuration())
+        let destination = try makeDatabasePool(filename: "destination.sqlite", configuration: Configuration())
         
         try source.write { db in
             try db.execute(sql: "CREATE TABLE items (id INTEGER PRIMARY KEY)")