+++ Tests/GRDBTests/EncryptionTests.swift
@@ -6,8 +6,11 @@ class EncryptionTests: GRDBTestCase {
     
     func testDatabaseQueueWithPassphraseToDatabaseQueueWithPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -15,8 +18,11 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
             }
@@ -25,8 +31,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabaseQueueWithPassphraseToDatabaseQueueWithoutPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -34,9 +43,8 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = nil
             do {
-                _ = try makeDatabaseQueue(filename: "test.sqlite")
+                _ = try makeDatabaseQueue(filename: "test.sqlite", configuration: Configuration())
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
@@ -47,8 +55,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabaseQueueWithPassphraseToDatabaseQueueWithWrongPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -56,9 +67,12 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "wrong"
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("wrong")
+            }
             do {
-                _ = try makeDatabaseQueue(filename: "test.sqlite")
+                _ = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
@@ -66,11 +80,14 @@ class EncryptionTests: GRDBTestCase {
             }
         }
     }
-
+    
     func testDatabaseQueueWithPassphraseToDatabaseQueueWithNewPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -78,9 +95,17 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
-            try dbQueue.change(passphrase: "newSecret")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
+            try dbQueue.inDatabase { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
+            }
+            try dbQueue.write { db in
+                try db.changePassphrase("newSecret")
+            }
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "INSERT INTO data (value) VALUES (2)")
             }
@@ -90,8 +115,11 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "newSecret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("newSecret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
             }
@@ -100,8 +128,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabaseQueueWithPassphraseToDatabasePoolWithPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -109,8 +140,11 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
             try dbPool.read { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
             }
@@ -119,8 +153,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabaseQueueWithPassphraseToDatabasePoolWithoutPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -128,9 +165,8 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = nil
             do {
-                _ = try makeDatabasePool(filename: "test.sqlite")
+                _ = try makeDatabasePool(filename: "test.sqlite", configuration: Configuration())
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
@@ -141,8 +177,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabaseQueueWithPassphraseToDatabasePoolWithWrongPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -150,9 +189,12 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "wrong"
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("wrong")
+            }
             do {
-                _ = try makeDatabasePool(filename: "test.sqlite")
+                _ = try makeDatabasePool(filename: "test.sqlite", configuration: config)
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
@@ -163,8 +205,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabaseQueueWithPassphraseToDatabasePoolWithNewPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -172,18 +217,35 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
-            try dbPool.change(passphrase: "newSecret")
+            var passphrase = "secret"
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase(passphrase)
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
+            try dbPool.read { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
+            }
+            try dbPool.barrierWriteWithoutTransaction { db in
+                passphrase = "newSecret"
+                try db.changePassphrase(passphrase)
+                dbPool.invalidateReadOnlyConnections()
+            }
             try dbPool.write { db in
                 try db.execute(sql: "INSERT INTO data (value) VALUES (2)")
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+            try dbPool.read { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
             }
         }
         
         do {
-            dbConfiguration.passphrase = "newSecret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("newSecret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
             try dbPool.read { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
             }
@@ -192,8 +254,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabasePoolWithPassphraseToDatabasePoolWithPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
             try dbPool.write { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -201,8 +266,11 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
             try dbPool.read { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
             }
@@ -211,8 +279,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabasePoolWithPassphraseToDatabasePoolWithoutPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
             try dbPool.write { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -220,9 +291,8 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = nil
             do {
-                _ = try makeDatabasePool(filename: "test.sqlite")
+                _ = try makeDatabasePool(filename: "test.sqlite", configuration: Configuration())
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
@@ -233,8 +303,11 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabasePoolWithPassphraseToDatabasePoolWithWrongPassphrase() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
             try dbPool.write { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -242,9 +315,12 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "wrong"
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("wrong")
+            }
             do {
-                _ = try makeDatabasePool(filename: "test.sqlite")
+                _ = try makeDatabasePool(filename: "test.sqlite", configuration: config)
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
@@ -256,8 +332,11 @@ class EncryptionTests: GRDBTestCase {
     func testDatabasePoolWithPassphraseToDatabasePoolWithNewPassphrase() throws {
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
             try dbPool.write { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
@@ -265,18 +344,35 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
-            try dbPool.change(passphrase: "newSecret")
+            var passphrase = "secret"
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase(passphrase)
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
+            try dbPool.read { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
+            }
+            try dbPool.barrierWriteWithoutTransaction { db in
+                passphrase = "newSecret"
+                try db.changePassphrase(passphrase)
+                dbPool.invalidateReadOnlyConnections()
+            }
             try dbPool.write { db in
                 try db.execute(sql: "INSERT INTO data (value) VALUES (2)")
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+            try dbPool.read { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
             }
         }
         
         do {
-            dbConfiguration.passphrase = "newSecret"
-            let dbPool = try makeDatabasePool(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("newSecret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
             try dbPool.read { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
             }
@@ -285,8 +381,7 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabaseQueueWithPragmaPassphraseToDatabaseQueueWithPassphrase() throws {
         do {
-            dbConfiguration.passphrase = nil
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: Configuration())
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "PRAGMA key = 'secret'")
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
@@ -295,8 +390,11 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
             }
@@ -305,8 +403,7 @@ class EncryptionTests: GRDBTestCase {
 
     func testDatabaseQueueWithPragmaPassphraseToDatabaseQueueWithoutPassphrase() throws {
         do {
-            dbConfiguration.passphrase = nil
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: Configuration())
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "PRAGMA key = 'secret'")
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
@@ -315,9 +412,8 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = nil
             do {
-                _ = try makeDatabaseQueue(filename: "test.sqlite")
+                _ = try makeDatabaseQueue(filename: "test.sqlite", configuration: Configuration())
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
@@ -328,24 +424,26 @@ class EncryptionTests: GRDBTestCase {
     
     func testCipherPageSize() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            dbConfiguration.prepareDatabase = { db in
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
                 try db.execute(sql: "PRAGMA cipher_page_size = 8192")
             }
             
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase({ db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "PRAGMA cipher_page_size")!, 8192)
             })
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            dbConfiguration.prepareDatabase = { db in
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
                 try db.execute(sql: "PRAGMA cipher_page_size = 4096")
             }
             
-            let dbPool = try makeDatabasePool(filename: "testpool.sqlite")
+            let dbPool = try makeDatabasePool(filename: "testpool.sqlite", configuration: config)
             try dbPool.write({ db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "PRAGMA cipher_page_size")!, 4096)
                 try db.execute(sql: "CREATE TABLE data(value INTEGER)")
@@ -361,24 +459,26 @@ class EncryptionTests: GRDBTestCase {
     
     func testCipherKDFIterations() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            dbConfiguration.prepareDatabase = { db in
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
                 try db.execute(sql: "PRAGMA kdf_iter = 128000")
             }
             
-            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "PRAGMA kdf_iter"), 128000)
             }
         }
 
         do {
-            dbConfiguration.passphrase = "secret"
-            dbConfiguration.prepareDatabase = { db in
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
                 try db.execute(sql: "PRAGMA kdf_iter = 128000")
             }
 
-            let dbPool = try makeDatabasePool(filename: "testpool.sqlite")
+            let dbPool = try makeDatabasePool(filename: "testpool.sqlite", configuration: config)
             try dbPool.write { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "PRAGMA kdf_iter"), 128000)
                 try db.execute(sql: "CREATE TABLE data(value INTEGER)")
@@ -394,12 +494,13 @@ class EncryptionTests: GRDBTestCase {
 
     func testCipherWithMismatchedKDFIterations() throws {
         do {
-            dbConfiguration.passphrase = "secret"
-            dbConfiguration.prepareDatabase = { db in
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
                 try db.execute(sql: "PRAGMA kdf_iter = 128000")
             }
 
-            let dbPool = try makeDatabasePool(filename: "testpool.sqlite")
+            let dbPool = try makeDatabasePool(filename: "testpool.sqlite", configuration: config)
             try dbPool.write { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "PRAGMA kdf_iter"), 128000)
                 try db.execute(sql: "CREATE TABLE data(value INTEGER)")
@@ -413,13 +514,14 @@ class EncryptionTests: GRDBTestCase {
         }
 
         do {
-            dbConfiguration.passphrase = "secret"
-            dbConfiguration.prepareDatabase = { db in
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
                 try db.execute(sql: "PRAGMA kdf_iter = 64000")
             }
 
             do {
-                let dbPool = try makeDatabasePool(filename: "testpool.sqlite")
+                let dbPool = try makeDatabasePool(filename: "testpool.sqlite", configuration: config)
 
                 try dbPool.read { db in
                     XCTAssertEqual(try Int.fetchOne(db, sql: "PRAGMA kdf_iter"), 64000)
@@ -436,29 +538,25 @@ class EncryptionTests: GRDBTestCase {
     func testExportPlainTextDatabaseToEncryptedDatabase() throws {
         // See https://discuss.zetetic.net/t/how-to-encrypt-a-plaintext-sqlite-database-to-use-sqlcipher-and-avoid-file-is-encrypted-or-is-not-a-database-errors/868?source_topic_id=939
         do {
-            // https://github.com/sqlcipher/sqlcipher/issues/216
-            // SQLCipher 3.4.1 crashes when sqlcipher_export() is called and a
-            // trace hook has been installed. So we disable query tracing for
-            // this test.
-            dbConfiguration.trace = nil
-            
-            dbConfiguration.passphrase = nil
-            let plainTextDBQueue = try makeDatabaseQueue(filename: "plaintext.sqlite")
+            let plainTextDBQueue = try makeDatabaseQueue(filename: "plaintext.sqlite", configuration: Configuration())
             try plainTextDBQueue.inDatabase { db in
                 try db.execute(sql: "CREATE TABLE data (value INTEGER)")
                 try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
             }
             
-            dbConfiguration.passphrase = "secret"
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
             do {
-                _ = try makeDatabaseQueue(filename: "plaintext.sqlite")
+                _ = try makeDatabaseQueue(filename: "plaintext.sqlite", configuration: config)
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
                 XCTAssertEqual(error.message!, "file is not a database")
             }
             
-            let encryptedDBQueue = try makeDatabaseQueue(filename: "encrypted.sqlite")
+            let encryptedDBQueue = try makeDatabaseQueue(filename: "encrypted.sqlite", configuration: config)
             
             try plainTextDBQueue.inDatabase { db in
                 try db.execute(sql: "ATTACH DATABASE ? AS encrypted KEY ?", arguments: [encryptedDBQueue.path, "secret"])
@@ -468,8 +566,11 @@ class EncryptionTests: GRDBTestCase {
         }
         
         do {
-            dbConfiguration.passphrase = "secret"
-            let dbQueue = try makeDatabaseQueue(filename: "encrypted.sqlite")
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "encrypted.sqlite", configuration: config)
             try dbQueue.inDatabase { db in
                 XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
             }
@@ -487,8 +588,8 @@ class EncryptionTests: GRDBTestCase {
             let testBundle = Bundle(for: type(of: self))
             let path = testBundle.url(forResource: "db", withExtension: "SQLCipher3")!.path
             var configuration = Configuration()
-            configuration.passphrase = "secret"
             configuration.prepareDatabase = { db in
+                try db.usePassphrase("secret")
                 try db.execute(sql: "PRAGMA cipher_compatibility = 3")
             }
             
@@ -502,6 +603,164 @@ class EncryptionTests: GRDBTestCase {
                 let dbPool = try DatabasePool(path: path, configuration: configuration)
                 let success = try dbPool.read { try String.fetchOne($0, sql: "SELECT a FROM t") }
                 XCTAssertEqual(success, "success")
+            }
+        }
+    }
+    
+    // MARK: - Deprecated
+    
+    func testDeprecatedPassphrase() throws {
+        do {
+            var config = Configuration()
+            config.passphrase = "secret"
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
+            try dbQueue.inDatabase { db in
+                try db.execute(sql: "CREATE TABLE data (value INTEGER)")
+                try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
+            }
+        }
+
+        do {
+            var config = Configuration()
+            config.passphrase = "secret"
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
+            try dbQueue.inDatabase { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 1)
+            }
+        }
+
+        do {
+            var config = Configuration()
+            config.passphrase = nil
+            do {
+                _ = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
+                XCTFail("Expected error")
+            } catch let error as DatabaseError {
+                XCTAssertEqual(error.resultCode, .SQLITE_NOTADB)
+                XCTAssertEqual(error.message!, "file is not a database")
+            }
+        }
+    }
+    
+    func testDeprecatedDatabaseQueueWithPassphraseToDatabaseQueueWithNewPassphrase() throws {
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
+            try dbQueue.inDatabase { db in
+                try db.execute(sql: "CREATE TABLE data (value INTEGER)")
+                try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
+            }
+        }
+        
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
+            try dbQueue.change(passphrase: "newSecret")
+            try dbQueue.inDatabase { db in
+                try db.execute(sql: "INSERT INTO data (value) VALUES (2)")
+            }
+            try dbQueue.inDatabase { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+        }
+        
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("newSecret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
+            try dbQueue.inDatabase { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+        }
+    }
+    
+    func testDeprecatedDatabaseQueueWithPassphraseToDatabasePoolWithNewPassphrase() throws {
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbQueue = try makeDatabaseQueue(filename: "test.sqlite", configuration: config)
+            try dbQueue.inDatabase { db in
+                try db.execute(sql: "CREATE TABLE data (value INTEGER)")
+                try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
+            }
+        }
+        
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
+            try dbPool.change(passphrase: "newSecret")
+            try dbPool.write { db in
+                try db.execute(sql: "INSERT INTO data (value) VALUES (2)")
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+            try dbPool.read { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+        }
+        
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("newSecret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
+            try dbPool.read { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+        }
+    }
+
+    func testDeprecatedDatabasePoolWithPassphraseToDatabasePoolWithNewPassphrase() throws {
+        
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
+            try dbPool.write { db in
+                try db.execute(sql: "CREATE TABLE data (value INTEGER)")
+                try db.execute(sql: "INSERT INTO data (value) VALUES (1)")
+            }
+        }
+        
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("secret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
+            try dbPool.change(passphrase: "newSecret")
+            try dbPool.write { db in
+                try db.execute(sql: "INSERT INTO data (value) VALUES (2)")
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+            try dbPool.read { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
+            }
+        }
+        
+        do {
+            var config = Configuration()
+            config.prepareDatabase = { db in
+                try db.usePassphrase("newSecret")
+            }
+            let dbPool = try makeDatabasePool(filename: "test.sqlite", configuration: config)
+            try dbPool.read { db in
+                XCTAssertEqual(try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM data")!, 2)
             }
         }
     }